+++ src/app/components/widgets/base/widget/widget.component.ts
@@ -26,6 +26,7 @@ import {ModalService} from '../../../../services/modal.service';
 import {TextAreaComponent} from '../../../ui/text-area/text-area.component';
 import {BaseChartClass} from '../../charts/base-chart.class';
 import {ActivatedRoute} from '@angular/router';
+import {ShareDashboardComponent} from "../../../ui/share-dashboard/share-dashboard/share-dashboard.component";
 
 @Component({
     selector: 'dsw-widget',
@@ -289,14 +290,23 @@ export class WidgetComponent implements OnInit, OnDestroy, AfterViewInit {
             return;
         }
         const mdx = this.component.getMDX();
-        this.ms.show({
-            title: 'Copy MDX',
-            component: TextAreaComponent,
+
+        const copyModal = {
+            title: '',
+            component: ShareDashboardComponent,
             closeByEsc: true,
-            onComponentInit: (c: TextAreaComponent) => {
-                c.value = mdx;
+            buttons: [],
+            class: 'modal-no-border',
+            onComponentInit: (sd: ShareDashboardComponent) => {
+                sd.title = 'Copy MDX';
+                sd.btnTitle = 'Copy';
+                sd.shareUrl = mdx;
+                sd.onCopy = () => {
+                    this.ms.close(copyModal);
+                };
             }
-        });
+        };
+        this.ms.show(copyModal);
     }
 
     /**
@@ -362,14 +372,22 @@ export class WidgetComponent implements OnInit, OnDestroy, AfterViewInit {
         }
         html += '></iframe>';
 
-        this.ms.show({
-            title: 'Share Widget',
-            component: TextAreaComponent,
+        const shareModal = {
+            title: '',
+            component: ShareDashboardComponent,
             closeByEsc: true,
-            onComponentInit: (c: TextAreaComponent) => {
-                c.value = html;
+            buttons: [],
+            class: 'modal-no-border',
+            onComponentInit: (sd: ShareDashboardComponent) => {
+                sd.title = 'Share widget';
+                sd.btnTitle = 'Copy';
+                sd.shareUrl = html;
+                sd.onCopy = () => {
+                    this.ms.close(shareModal);
+                };
             }
-        });
+        };
+        this.ms.show(shareModal);
 
     }