+++ src/app/components/widgets/text/wtext.component.ts
@@ -75,6 +75,11 @@ export class WTextComponent extends BaseWidget implements OnInit, AfterViewInit 
         this.hideLoading();
         if (result) {
             for (let i = 0; i < result.Cols[0].tuples.length; i++) {
+                if (this.widget.dataProperties) {
+                    if (!this.getDataPropByDataValue(result.Cols[0].tuples[i].dimension)) {
+                        continue;
+                    }
+                }
                 let propFmt = '';
                 if (this.widget && this.widget.properties && this.widget.properties.format) {
                     propFmt = this.widget.properties.format;