+++ src/app/services/util.service.ts
@@ -95,4 +95,27 @@ export class UtilService {
     isMobile(): boolean {
         return document.body.offsetWidth <= 576;
     }
+
+    copyToClipboard(text: string) {
+        const clipboardData = window['clipboardData'];
+        if (clipboardData && clipboardData.setData) {
+            // Internet Explorer-specific code path to prevent textarea being shown while dialog is visible.
+            return clipboardData.setData('Text', text);
+
+        } else if (document.queryCommandSupported && document.queryCommandSupported('copy')) {
+            const textarea = document.createElement('textarea');
+            textarea.textContent = text;
+            textarea.style.position = 'fixed';  // Prevent scrolling to bottom of page in Microsoft Edge.
+            document.body.appendChild(textarea);
+            textarea.select();
+            try {
+                return document.execCommand('copy');  // Security exception may be thrown by some browsers.
+            } catch (ex) {
+                console.warn('Copy to clipboard failed.', ex);
+                return false;
+            } finally {
+                document.body.removeChild(textarea);
+            }
+        }
+    }
 }