+++ src/environments/dsw.ts
@@ -1,4 +1,4 @@
-import * as pkg from '../../package.json';
+import pkg from '../../package.json';
 
 export const dsw = {
     // TODO: add type for addons